+++ src/collective/recaptcha/settings.py
@@ -1,37 +1,18 @@
 # coding=utf-8
 from collective.recaptcha import RecaptchaMessageFactory as _
-from persistent import Persistent
+from plone.app.registry.browser.controlpanel import RegistryEditForm
 from plone.registry.interfaces import IRegistry
 from zope import schema
-from zope.annotation import factory
-from zope.annotation import IAttributeAnnotatable
-from zope.component import adapter
 from zope.component import getUtility
-from zope.component.hooks import getSite
-from zope.interface import implementer
 from zope.interface import Interface
 
 
 try:
-    from zope.formlib.form import FormFields
-except ImportError:
-    # formlib missing (Plone 5?)
-    FormFields = None
-
-
-try:
-    # formlib missing (Plone 5?)
-    from plone.app.registry.browser.controlpanel import RegistryEditForm as EditForm
-except ImportError:
-    # Zope 2.12+
-    from five.formlib.formbase import EditForm
-
-try:
     from plone.formwidget.recaptcha.interfaces import IReCaptchaSettings
 
-    TRY_REGISTRY = True
+    TRY_FORMWIDGET = True
 except ImportError:
-    TRY_REGISTRY = False
+    TRY_FORMWIDGET = False
 
 
 class IRecaptchaSettings(Interface):
@@ -41,20 +22,9 @@ class IRecaptchaSettings(Interface):
     private_key = schema.TextLine(title=_(u"Secret Key"))
 
 
-@implementer(IRecaptchaSettings)
-@adapter(IAttributeAnnotatable)
-class RecaptchaSettingsAnnotations(Persistent):
-    def __init__(self):
-        self.public_key = None
-        self.private_key = None
-
-
-RecaptchaSettings = factory(RecaptchaSettingsAnnotations)
-
-
 def getRecaptchaSettings():
     registry = getUtility(IRegistry)
-    if TRY_REGISTRY:
+    if TRY_FORMWIDGET:
         # if plone.formwidget.recaptcha is installed, try getting
         # its settings from the registry
         try:
@@ -63,22 +33,11 @@ def getRecaptchaSettings():
                 return settings
         except (AttributeError, KeyError):
             pass
-    # try getting settings from the registry first
-    try:
-        settings = registry.forInterface(IRecaptchaSettings)
-        if settings.public_key and settings.private_key:
-            return settings
-    except KeyError:
-        # fall back to our storage of an annotation on the site if the settings
-        # haven't been configured
-        site = getSite()
-        return IRecaptchaSettings(site)
+    settings = registry.forInterface(IRecaptchaSettings)
+    if settings.public_key and settings.private_key:
+        return settings
 
 
-class RecaptchaSettingsForm(EditForm):
+class RecaptchaSettingsForm(RegistryEditForm):
     schema = IRecaptchaSettings
     label = _(u"Recaptcha settings")
-
-    if FormFields:
-        # formlib missing (Plone 5?)
-        form_fields = FormFields(IRecaptchaSettings)